Refactor session routes into their own file
